+++ NEWS.md
@@ -1,3 +1,7 @@
+# asleep 0.2.0
+
+* Added `py_asleep` for python calling and execution inside environment.
+
 # asleep 0.1.0
 
 * Initial submission.